Use url instead of uri
My reasoning is this: the more commonly used term is url. uri is easily
mistaken for url, since i, I and l all look very similar. To avoid
confusion, use the colloquialism.

If we want to be picky, I think technically urls are uris that define
the network location of a resource, so I suppose url is also more
correct, fwiw.